fix(2a): confidence as decimal(5,4), simplify schema test setup

+++ backend/database/migrations/2026_05_25_000000_create_document_extractions_table.php
@@ -17,7 +17,7 @@ return new class extends Migration
             $table->integer('char_count')->nullable();
             $table->string('ocr_engine')->nullable();
             $table->string('extractor_type')->nullable();
-            $table->float('confidence')->nullable();
+            $table->decimal('confidence', 5, 4)->nullable();
             $table->timestamp('extracted_at')->nullable();
             $table->timestamps();
         });
